Refactoring & minor improvements
