+++ src/main/scala/akka/MessageWrapper.scala
@@ -1,0 +1,3 @@
+package akka
+
+case class MessageWrapper(id: Int, msg: AnyRef)
